// api/model/subscriberDto.ts
/**
 * Student-Management-System-API
 * The Student-Management-Sytem-API. <a href='http://localhost:3000/api-json'>JSON</a>
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SubscriberDto { 
    /**
     * Name of your application.
     */
    name: string;
    /**
     * Complete URL to an endpoint in your application that receives HTTP-POST messages.
     */
    url: string;
    /**
     * Map of events that you want to be notified about. Refer to the Event-Enum for event names. You can use 'ALL' to subscribe to all events.
     */
    events: any;
    /**
     * Date of the creation or latest update to the subscription.
     */
    updateDate?: Date;
}